Move block-status API under /api
