+++ season_pass/api/__init__.py
@@ -1,12 +1,15 @@
+import os
 from datetime import datetime, timezone, timedelta
 
+import requests
 from fastapi import APIRouter, Depends
-from sqlalchemy import select, or_
+from sqlalchemy import select, or_, func
 from sqlalchemy.orm import Session
 from starlette.responses import JSONResponse
 
 from dependencies import session
 from enums import TxStatus
+from models.action import Block
 from models.user import Claim
 from season_pass import settings
 from season_pass.api import season_pass, user, tmp
@@ -26,6 +29,39 @@ if settings.stage != "mainnet":
 
 for view in __all__:
     router.include_router(view.router)
+
+
+@router.get("/block-status")
+def block_status(sess=Depends(session)):
+    resp = requests.post(
+        os.environ["ODIN_VALIDATOR_URL"],
+        json={"query": "{ nodeStatus { tip { index } } }"}
+    )
+    odin_tip = resp.json()["data"]["nodeStatus"]["tip"]["index"]
+    resp = requests.post(
+        os.environ["HEIMDALL_VALIDATOR_URL"],
+        json={"query": "{ nodeStatus { tip { index } } }"}
+    )
+    heimdall_tip = resp.json()["data"]["nodeStatus"]["tip"]["index"]
+
+    latest = (sess.query(Block.planet_id, func.max(Block.index))
+              .group_by(Block.planet_id).order_by(Block.planet_id)
+              ).all()
+
+    err = abs(latest[0][1] - odin_tip) > 10 or abs(latest[1][1] - heimdall_tip) > 10
+    msg = {
+        latest[0][0].decode(): {
+            "headless_tip": odin_tip,
+            "db_tip": latest[0][1],
+            "diverge": abs(odin_tip - latest[0][1]),
+        },
+        latest[1][0].decode(): {
+            "headless_tip": heimdall_tip,
+            "db_tip": latest[1][1],
+            "diverge": abs(heimdall_tip - latest[1][1]),
+        },
+    }
+    return JSONResponse(status_code=503 if err else 200, content=msg, )
 
 
 @router.get("/invalid-claim")
